chore: fixed 'pfverts' line element visit to comply with the convention that the neighbors of line elements are opposite to nodes chore: removed old references to 'super_group' chore: improved printLineElementRegion chore: cleaned up some for loops that required explicit specification of loop variable

+++ source/modules/integration/pde_integrators/PDE_Integrator.cpp
@@ -1360,7 +1360,7 @@ void PDE_Integrator<dim,COMPUTATION_DOMAIN>::Accumulate( const COMPUTATION_DOMAI
               it_lhs=this->thread_lhs_operators_[tid].begin(); it_lhs!=this->thread_lhs_operators_[tid].end(); it_lhs++ ){
             if ( !(*it_lhs).second->AddLater() && !(*it_lhs).second->SubtractLater() ){
 #pragma omp for
-                for ( int32_t e = 0 ; e < gref.Elements(); e++ )
+                for ( size_t e = 0 ; e < gref.Elements(); e++ )
                 {
 //                    cout<<"element: "<<e<<endl;
 // TODO: check what is going on here with the parallel accumulation
@@ -1396,7 +1396,7 @@ void PDE_Integrator<dim,COMPUTATION_DOMAIN>::Accumulate( const COMPUTATION_DOMAI
               it_rhs=this->thread_rhs_operators_[tid].begin(); it_rhs!=this->thread_rhs_operators_[tid].end(); it_rhs++ ){
             if ( !(*it_rhs).second->AddLater() && !(*it_rhs).second->SubtractLater() ){
 #pragma omp for
-                for ( int32_t e = 0 ; e < gref.Elements(); e++ )
+                for ( size_t e = 0 ; e < gref.Elements(); e++ )
                 {
                     typename COMPUTATION_DOMAIN<dim>::CellType* eit = gref.E(e);
                     fe_tmp=eit->FE(); //save old pointer.
